improvement: move handlers out of routes
