+++ app/api/listings/handlers.ts
@@ -1,0 +1,346 @@
+import { and, asc, desc, eq, ilike, inArray, lte, or, sql } from "drizzle-orm";
+import { TypedNextResponse } from "next-rest-framework";
+
+import { db } from "@/db";
+import { listingImages, listings, properties, type ListingStatus } from "@/db/schema";
+import { getOptionalSession, requireListingWriteSession } from "@/lib/auth/session";
+import {
+  buildListingCustomFields,
+  centsToDollars,
+  DEFAULT_PROPERTY_COUNTRY,
+  dollarsToCents,
+  formatListingAddress,
+  formatListingTimeAgo,
+  getListingCoordinates,
+  getListingSquareFeet,
+  resolveListingStatusTimestamps,
+} from "@/lib/listings/store";
+import type {
+  CreateListingInput,
+  CreateListingResponse,
+  ListingListResponse,
+} from "@/shared/schemas/listings";
+
+export async function getListingsHandler(request: Request) {
+  const { searchParams } = new URL(request.url);
+
+  const page = Number(searchParams.get("page") ?? 1);
+  const limit = Number(searchParams.get("limit") ?? 20);
+  const status = searchParams.get("status");
+  const neighborhood = searchParams.get("neighborhood");
+  const bedrooms = searchParams.get("bedrooms");
+  const maxRent = searchParams.get("maxRent");
+  const accessibility = searchParams.get("accessibility") ?? undefined;
+  const search = searchParams.get("search");
+
+  const optionalSession =
+    status === "draft" || status === "archived"
+      ? await getOptionalSession()
+      : { session: null, authzUser: null };
+
+  const visibility = getListingListVisibility(optionalSession, status);
+
+  if (!visibility.isAccessible) {
+    return TypedNextResponse.json<ListingListResponse, 200, "application/json">({
+      data: [],
+      pagination: {
+        page,
+        limit,
+        total: 0,
+        totalPages: 0,
+      },
+    });
+  }
+
+  const filters = [eq(listings.status, visibility.status)];
+
+  if (visibility.ownerUserId) {
+    filters.push(eq(properties.ownerUserId, visibility.ownerUserId));
+  }
+
+  if (neighborhood) {
+    filters.push(ilike(properties.neighborhood, `%${neighborhood}%`));
+  }
+
+  if (bedrooms) {
+    filters.push(eq(listings.bedrooms, Number(bedrooms)));
+  }
+
+  if (maxRent) {
+    filters.push(lte(listings.monthlyRentCents, dollarsStringToCents(maxRent)));
+  }
+
+  const accessibilityFilter = buildAccessibilityFilter(
+    accessibility === "true" || accessibility === "false" ? accessibility : undefined,
+  );
+
+  if (accessibilityFilter) {
+    filters.push(accessibilityFilter);
+  }
+
+  if (search) {
+    const searchTerm = `%${search}%`;
+
+    filters.push(
+      or(
+        ilike(listings.title, searchTerm),
+        ilike(listings.description, searchTerm),
+        ilike(properties.name, searchTerm),
+        ilike(properties.street1, searchTerm),
+        ilike(properties.city, searchTerm),
+      )!,
+    );
+  }
+
+  const whereClause = and(...filters);
+  const offset = (page - 1) * limit;
+
+  const totalRows = await db
+    .select({ total: sql<number>`count(*)::int` })
+    .from(listings)
+    .innerJoin(properties, eq(listings.propertyId, properties.id))
+    .where(whereClause);
+  const total = Number(totalRows[0]?.total ?? 0);
+
+  const pagedRows = await db
+    .select({
+      id: listings.id,
+      monthlyRentCents: listings.monthlyRentCents,
+      bedrooms: listings.bedrooms,
+      bathrooms: listings.bathrooms,
+      squareFeet: listings.squareFeet,
+      customFields: listings.customFields,
+      unitNumber: listings.unitNumber,
+      publishedAt: listings.publishedAt,
+      createdAt: listings.createdAt,
+      street1: properties.street1,
+      city: properties.city,
+      latitude: properties.latitude,
+      longitude: properties.longitude,
+    })
+    .from(listings)
+    .innerJoin(properties, eq(listings.propertyId, properties.id))
+    .where(whereClause)
+    .orderBy(desc(listings.publishedAt), desc(listings.createdAt))
+    .limit(limit)
+    .offset(offset);
+
+  const listingIds = pagedRows.map((row) => row.id);
+  const imageRows =
+    listingIds.length > 0
+      ? await db
+          .select({
+            listingId: listingImages.listingId,
+            imageUrl: listingImages.imageUrl,
+            sortOrder: listingImages.sortOrder,
+          })
+          .from(listingImages)
+          .where(inArray(listingImages.listingId, listingIds))
+          .orderBy(asc(listingImages.sortOrder))
+      : [];
+
+  const imageByListingId = new Map<string, string>();
+  for (const image of imageRows) {
+    if (!imageByListingId.has(image.listingId)) {
+      imageByListingId.set(image.listingId, image.imageUrl);
+    }
+  }
+
+  return TypedNextResponse.json<ListingListResponse, 200, "application/json">({
+    data: pagedRows.map((row) => {
+      const coordinates = getListingCoordinates(row.latitude, row.longitude);
+      const listingSummary = {
+        id: row.id,
+        price: centsToDollars(row.monthlyRentCents),
+        address: formatListingAddress(row.street1, row.unitNumber),
+        city: row.city,
+        beds: row.bedrooms,
+        baths: row.bathrooms,
+        sqft: getListingSquareFeet(row.squareFeet, row.customFields),
+        imageUrl: imageByListingId.get(row.id),
+        timeAgo: formatListingTimeAgo(row.publishedAt, row.createdAt),
+      };
+
+      if (!coordinates) {
+        return listingSummary;
+      }
+
+      return {
+        ...listingSummary,
+        lat: coordinates.lat,
+        lng: coordinates.lng,
+      };
+    }),
+    pagination: {
+      page,
+      limit,
+      total,
+      totalPages: total === 0 ? 0 : Math.ceil(total / limit),
+    },
+  });
+}
+
+export async function createListingHandler(request: Request) {
+  const sessionResult = await requireListingWriteSession();
+
+  if (sessionResult.response) {
+    return sessionResult.response;
+  }
+
+  const { session } = sessionResult;
+  const body = (await request.json()) as CreateListingInput;
+  const primaryUnit = body.units[0];
+
+  if (!primaryUnit) {
+    return TypedNextResponse.json<{ message: string }, 400, "application/json">(
+      { message: "At least one unit is required" },
+      { status: 400 },
+    );
+  }
+
+  const statusTimestamps = resolveListingStatusTimestamps(body.status);
+  const customFields = buildListingCustomFields(body);
+
+  const createdListing = await db.transaction(async (tx) => {
+    const [property] = await tx
+      .insert(properties)
+      .values({
+        ownerUserId: session.user.id,
+        name: body.name,
+        street1: body.address.street,
+        street2: null,
+        city: body.address.city,
+        province: body.address.province,
+        postalCode: body.address.postalCode,
+        country: DEFAULT_PROPERTY_COUNTRY,
+        neighborhood: body.address.neighborhood ?? null,
+        latitude: body.address.latitude ?? null,
+        longitude: body.address.longitude ?? null,
+        contactName: body.contact.name,
+        contactEmail: body.contact.email,
+        contactPhone: body.contact.phone,
+        createdByUserId: session.user.id,
+        updatedByUserId: session.user.id,
+      })
+      .returning({ id: properties.id });
+
+    if (!property) {
+      throw new Error("Failed to create property.");
+    }
+
+    const [listing] = await tx
+      .insert(listings)
+      .values({
+        propertyId: property.id,
+        createdByUserId: session.user.id,
+        updatedByUserId: session.user.id,
+        title: body.name,
+        description: body.description,
+        status: body.status,
+        unitNumber: null,
+        bedrooms: primaryUnit.bedrooms,
+        bathrooms: primaryUnit.bathrooms,
+        squareFeet: primaryUnit.sqft,
+        monthlyRentCents: primaryUnit.rent * 100,
+        availableOn: primaryUnit.availableDate,
+        maxIncomeCents: dollarsToCents(body.eligibilityCriteria.maxIncome),
+        applicationUrl:
+          body.applicationMethod === "external_link" ? (body.externalApplicationUrl ?? null) : null,
+        applicationEmail: body.contact.email,
+        applicationPhone: body.contact.phone,
+        applicationInstructions: null,
+        customFields,
+        publishedAt: statusTimestamps.publishedAt,
+        archivedAt: statusTimestamps.archivedAt,
+      })
+      .returning({ id: listings.id });
+
+    if (!listing) {
+      throw new Error("Failed to create listing.");
+    }
+
+    if (body.images.length > 0) {
+      await tx.insert(listingImages).values(
+        body.images.map((imageUrl, index) => ({
+          listingId: listing.id,
+          imageUrl,
+          altText: `${body.name} image ${index + 1}`,
+          sortOrder: index,
+        })),
+      );
+    }
+
+    return listing;
+  });
+
+  return TypedNextResponse.json<CreateListingResponse, 201, "application/json">(
+    { message: "Listing created", data: { id: createdListing.id, ...body } },
+    { status: 201 },
+  );
+}
+
+function getListingListVisibility(
+  sessionResult: Awaited<ReturnType<typeof getOptionalSession>>,
+  requestedStatus: string | null,
+) {
+  type ListingVisibility = {
+    status: ListingStatus;
+    ownerUserId: string | null;
+    isAccessible: boolean;
+  };
+
+  if (requestedStatus === "draft" || requestedStatus === "archived") {
+    if (sessionResult.authzUser?.role === "admin") {
+      return {
+        status: requestedStatus,
+        ownerUserId: null,
+        isAccessible: true,
+      } satisfies ListingVisibility;
+    }
+
+    if (sessionResult.authzUser?.role === "partner" && sessionResult.session) {
+      return {
+        status: requestedStatus,
+        ownerUserId: sessionResult.session.user.id,
+        isAccessible: true,
+      } satisfies ListingVisibility;
+    }
+
+    return {
+      status: "published",
+      ownerUserId: null,
+      isAccessible: false,
+    } satisfies ListingVisibility;
+  }
+
+  return {
+    status: "published",
+    ownerUserId: null,
+    isAccessible: true,
+  } satisfies ListingVisibility;
+}
+
+function dollarsStringToCents(value: string) {
+  const [whole, fractional = ""] = value.split(".");
+  const cents = `${fractional}00`.slice(0, 2);
+
+  return Number(whole) * 100 + Number(cents);
+}
+
+function buildAccessibilityFilter(accessibility: "true" | "false" | undefined) {
+  if (!accessibility) {
+    return undefined;
+  }
+
+  const hasAccessibility = sql<boolean>`
+    jsonb_array_length(
+      case
+        when jsonb_typeof(${listings.customFields} -> 'accessibilityFeatures') = 'array'
+          then ${listings.customFields} -> 'accessibilityFeatures'
+        else '[]'::jsonb
+      end
+    ) > 0
+  `;
+
+  return accessibility === "true" ? hasAccessibility : sql<boolean>`not (${hasAccessibility})`;
+}
